# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Valuation Engine API",
    "description": "Financial Valuation Platform with DCF, DuPont, and Comps analysis",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/session/{session_id}/export-pdf": {
      "post": {
        "tags": [
          "PDF Export"
        ],
        "summary": "Export Pdf",
        "description": "Generate a PDF investment memorandum from valuation results.\n\nWorkflow:\n1. Retrieve Step 9 confirmed outputs and Step 10 results from session\n2. Build HTML template with all valuation data\n3. Convert to PDF using ReportLab\n4. Return binary PDF stream with Content-Disposition header\n\nReturns:\n    StreamingResponse with PDF binary data",
        "operationId": "export_pdf_api_session__session_id__export_pdf_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPdfRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportPdfRequest": {
        "properties": {
          "methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Methods",
            "description": "Valuation methods to include in report",
            "default": [
              "DCF",
              "DUPONT",
              "COMPS"
            ]
          },
          "market": {
            "type": "string",
            "title": "Market",
            "description": "Market type",
            "default": "international"
          },
          "include_sensitivity": {
            "type": "boolean",
            "title": "Include Sensitivity",
            "description": "Include sensitivity analysis tables",
            "default": true
          },
          "include_peer_comparison": {
            "type": "boolean",
            "title": "Include Peer Comparison",
            "description": "Include peer comparison data",
            "default": true
          }
        },
        "type": "object",
        "title": "ExportPdfRequest",
        "description": "Request to generate PDF export"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}